feat: remove legacy radix router implementation and consolidate routing logic into a new optimized router with caching and performance enhancements
