--- src/routing/radix-router-optimized.ts
@@ -1,430 +1,0 @@
-/**
- * Highly Optimized Radix Router
- *
- * This implementation incorporates multiple optimizations:
- * - Path segment caching to avoid string parsing
- * - Fast path optimization for common routes
- * - Instance pooling for router nodes
- * - Bitmap-based node indexing for faster child lookup
- * - SIMD acceleration for pattern matching where available
- */
-
-import { IncomingMessage, ServerResponse } from 'node:http';
-import { HttpMethod } from '../http/http-method.js';
-import { MiddlewareHandler } from '../middleware/middleware.js';
-import { performance } from 'node:perf_hooks';
-
-// Define route handler type
-export type RouteHandler = (_req: IncomingMessage, _res: ServerResponse) => Promise<void>;
-
-// Define route type
-export interface RadixRoute {
-  method: HttpMethod;
-  handler: RouteHandler;
-  middlewares: MiddlewareHandler[];
-  controller: any;
-}
-
-// Define route match result
-export interface RadixRouteMatch {
-  route: RadixRoute;
-  params: Record<string, string>;
-}
-
-// Node types in radix tree
-const _STATIC = 0;
-const _PARAM = 1;
-const _WILDCARD = 2;
-
-// Memory-efficient node class with optimization for static nodes
-class RadixNode {
-  // Node properties
-  type: number = _STATIC;
-  segment: string = '';
-  paramName: string = '';
-  children: RadixNode[] = [];
-  routes: Map<HttpMethod, RadixRoute> = new Map();
-
-  // Fast path optimization
-  fastPaths: Map<string, RadixNode> = new Map();
-
-  // For static nodes, use a bitmap index for fast child lookup
-  staticChildrenIndex: Uint32Array | null = null;
-
-  // Reusable param objects for better memory efficiency
-  private static paramObjectPool: Record<string, string>[] = [];
-
-  // Node pool for reuse
-  private static nodePool: RadixNode[] = [];
-
-  // Get a node from the pool or create a new one
-  static getNode(): RadixNode {
-    if (this.nodePool.length > 0) {
-      const node = this.nodePool.pop()!;
-      node.reset();
-      return node;
-    }
-    return new RadixNode();
-  }
-
-  // Release a node back to the pool
-  static releaseNode(node: RadixNode): void {
-    if (this.nodePool.length < 1000) { // Limit pool size
-      this.nodePool.push(node);
-    }
-  }
-
-  // Reset node to initial state
-  reset(): void {
-    this.type = _STATIC;
-    this.segment = '';
-    this.paramName = '';
-    this.children = [];
-    this.routes.clear();
-    this.fastPaths.clear();
-    this.staticChildrenIndex = null;
-  }
-
-  // Get param object from pool
-  static getParamObject(): Record<string, string> {
-    if (this.paramObjectPool.length > 0) {
-      return this.paramObjectPool.pop()!;
-    }
-    return {};
-  }
-
-  // Release param object back to pool
-  static releaseParamObject(params: Record<string, string>): void {
-    if (this.paramObjectPool.length < 1000) { // Limit pool size
-      Object.keys(params).forEach(key => delete params[key]);
-      this.paramObjectPool.push(params);
-    }
-  }
-
-  // Add a route to the tree
-  insert(path: string, route: RadixRoute): void {
-    const startTime = performance.now();
-    const normalizedPath = this.normalizePath(path);
-    const segments = normalizedPath.split('/').filter(Boolean);
-
-    this.insertInternal(segments, route, 0);
-
-    // Build fast paths for common routes (direct children)
-    if (segments.length === 1 && this.type === _STATIC) {
-      this.fastPaths.set(`${route.method}:${segments[0]}`, this.children.find(c => c.segment === segments[0]) || this);
-    }
-
-    // Rebuild bitmap index for static children
-    this.rebuildStaticIndex();
-
-    const endTime = performance.now();
-    if (endTime - startTime > 5) {
-      console.warn(`Slow route insertion: ${path} took ${(endTime - startTime).toFixed(2)}ms`);
-    }
-  }
-
-  // Internal insertion method
-  private insertInternal(segments: string[], route: RadixRoute, index: number): void {
-    // If we reached the end of the path, store the route at this node
-    if (index === segments.length) {
-      this.routes.set(route.method, route);
-      return;
-    }
-
-    const segment = segments[index];
-
-    // Check if this is a parameter segment (:name)
-    if (segment.startsWith(':')) {
-      // Parameter node
-      const paramName = segment.slice(1);
-      let paramNode = this.children.find(child => child.type === _PARAM);
-
-      if (!paramNode) {
-        paramNode = RadixNode.getNode();
-        paramNode.type = _PARAM;
-        paramNode.paramName = paramName;
-        this.children.push(paramNode);
-      } else if (paramNode.paramName !== paramName) {
-        throw new Error(`Cannot use two different param names for the same path segment: ${paramNode.paramName} and ${paramName}`);
-      }
-
-      paramNode.insertInternal(segments, route, index + 1);
-      return;
-    }
-
-    // Check if this is a wildcard segment (*)
-    if (segment === '*') {
-      // Wildcard node
-      let wildcardNode = this.children.find(child => child.type === _WILDCARD);
-
-      if (!wildcardNode) {
-        wildcardNode = RadixNode.getNode();
-        wildcardNode.type = _WILDCARD;
-        this.children.push(wildcardNode);
-      }
-
-      wildcardNode.routes.set(route.method, route);
-      return;
-    }
-
-    // Static node
-    let staticNode = this.children.find(child =>
-      child.type === _STATIC && child.segment === segment
-    );
-
-    if (!staticNode) {
-      staticNode = RadixNode.getNode();
-      staticNode.type = _STATIC;
-      staticNode.segment = segment!;
-      this.children.push(staticNode);
-    }
-
-    staticNode.insertInternal(segments, route, index + 1);
-  }
-
-  // Rebuild the bitmap index for static children
-  private rebuildStaticIndex(): void {
-    const staticChildren = this.children.filter(child => child.type === _STATIC);
-
-    if (staticChildren.length > 5) { // Only use bitmap for sufficient number of children
-      // Create a 256-bit (32 byte) bitmap for ASCII chars
-      this.staticChildrenIndex = new Uint32Array(8); // 8 * 32 = 256 bits
-
-      // Set bits for the first character of each static child
-      for (const child of staticChildren) {
-        const charCode = child.segment.charCodeAt(0);
-        const index = Math.floor(charCode / 32);
-        const bit = charCode % 32;
-        this.staticChildrenIndex[index] |= (1 << bit);
-      }
-    } else {
-      this.staticChildrenIndex = null;
-    }
-  }
-
-  // Find a route in the tree
-  search(path: string, method: HttpMethod): RadixRouteMatch | null {
-    // Fast path check for common routes
-    const fastPathKey = `${method}:${path.replace(/^\//, '')}`;
-    const fastPathNode = this.fastPaths.get(fastPathKey);
-
-    if (fastPathNode) {
-      const route = fastPathNode.routes.get(method);
-      if (route) {
-        return {
-          route,
-          params: {}
-        };
-      }
-    }
-
-    // Normal path
-    const normalizedPath = this.normalizePath(path);
-    const segments = normalizedPath.split('/').filter(Boolean);
-    const params = RadixNode.getParamObject();
-
-    const route = this.searchInternal(segments, method, params, 0);
-
-    if (route) {
-      return { route, params };
-    }
-
-    // Release the params object back to the pool
-    RadixNode.releaseParamObject(params);
-    return null;
-  }
-
-  // Internal search method
-  private searchInternal(
-    segments: string[],
-    method: HttpMethod,
-    params: Record<string, string>,
-    index: number
-  ): RadixRoute | null {
-    // If we reached the end of the path, look for a route at this node
-    if (index === segments.length) {
-      return this.routes.get(method) || null;
-    }
-
-    const segment = segments[index];
-
-    // First, try to find a static child - this is the most common case
-    if (this.staticChildrenIndex) {
-      // Use bitmap index for faster lookup
-      const charCode = segment.charCodeAt(0);
-      const bitmapIndex = Math.floor(charCode / 32);
-      const bitPosition = charCode % 32;
-
-      // Check if this first character exists in our bitmap
-      if (this.staticChildrenIndex[bitmapIndex] & (1 << bitPosition)) {
-        // We have at least one child with this first character
-        for (const child of this.children) {
-          if (child.type === _STATIC && child.segment === segment) {
-            const route = child.searchInternal(segments, method, params, index + 1);
-            if (route) return route;
-            break;
-          }
-        }
-      }
-    } else {
-      // Linear search for small number of children
-      for (const child of this.children) {
-        if (child.type === _STATIC && child.segment === segment) {
-          const route = child.searchInternal(segments, method, params, index + 1);
-          if (route) return route;
-          break;
-        }
-      }
-    }
-
-    // If no static match, try parameter nodes
-    for (const child of this.children) {
-      if (child.type === _PARAM) {
-        params[child.paramName] = segment!;
-        const route = child.searchInternal(segments, method, params, index + 1);
-        if (route) return route;
-        delete params[child.paramName]; // Backtrack
-      }
-    }
-
-    // Finally, try wildcard nodes
-    for (const child of this.children) {
-      if (child.type === _WILDCARD) {
-        return child.routes.get(method) || null;
-      }
-    }
-
-    return null;
-  }
-
-  // Normalize a path
-  private normalizePath(path: string): string {
-    if (!path.startsWith('/')) {
-      path = '/' + path;
-    }
-
-    if (path.endsWith('/') && path.length > 1) {
-      path = path.slice(0, -1);
-    }
-
-    return path;
-  }
-}
-
-// Main router class
-export class OptimizedRadixRouter {
-  private root = new RadixNode();
-  private globalPrefix: string;
-  private routeCache: Map<string, RadixRouteMatch | null> = new Map();
-  private readonly CACHE_MAX_SIZE = 10000; // Maximum cache size
-
-  // Statistics for monitoring
-  private stats = {
-    hits: 0,
-    misses: 0,
-    inserts: 0,
-    searches: 0
-  };
-
-  constructor(globalPrefix: string = '') {
-    this.globalPrefix = this.normalizePath(globalPrefix);
-  }
-
-  // Add a route to the router
-  addRoute(
-    method: HttpMethod,
-    path: string,
-    handler: RouteHandler,
-    middlewares: MiddlewareHandler[] = [],
-    controller: any = null
-  ): void {
-    this.stats.inserts++;
-
-    // Combine global prefix with path
-    const fullPath = this.combinePaths(this.globalPrefix, path);
-
-    // Create route object
-    const route: RadixRoute = {
-      method,
-      handler,
-      middlewares,
-      controller
-    };
-
-    // Insert route into the tree
-    this.root.insert(fullPath, route);
-
-    // Invalidate cache entries that might be affected
-    for (const key of this.routeCache.keys()) {
-      if (key.endsWith(`:${fullPath}`)) {
-        this.routeCache.delete(key);
-      }
-    }
-  }
-
-  // Find a route for a request
-  findRoute(method: HttpMethod, path: string): RadixRouteMatch | null {
-    this.stats.searches++;
-
-    // Check cache first
-    const cacheKey = `${method}:${path}`;
-    if (this.routeCache.has(cacheKey)) {
-      this.stats.hits++;
-      return this.routeCache.get(cacheKey)!;
-    }
-
-    this.stats.misses++;
-
-    // Not in cache, perform tree search
-    const result = this.root.search(path, method);
-
-    // Cache the result (including null for not found)
-    if (this.routeCache.size >= this.CACHE_MAX_SIZE) {
-      // LRU-like eviction: remove oldest entries (first 10% of cache)
-      const keysToRemove = Math.max(1, Math.floor(this.CACHE_MAX_SIZE * 0.1));
-      const keys = [...this.routeCache.keys()].slice(0, keysToRemove);
-      keys.forEach(key => this.routeCache.delete(key));
-    }
-
-    this.routeCache.set(cacheKey, result);
-
-    return result;
-  }
-
-  // Get router statistics
-  getStats(): { hits: number, misses: number, inserts: number, searches: number, cacheSize: number } {
-    return {
-      ...this.stats,
-      cacheSize: this.routeCache.size
-    };
-  }
-
-  // Clear the route cache
-  clearCache(): void {
-    this.routeCache.clear();
-  }
-
-  // Normalize a path
-  private normalizePath(path: string): string {
-    let normalized = path;
-
-    if (!normalized.startsWith('/')) {
-      normalized = '/' + normalized;
-    }
-
-    if (normalized.endsWith('/') && normalized.length > 1) {
-      normalized = normalized.slice(0, -1);
-    }
-
-    return normalized;
-  }
-
-  // Combine multiple paths
-  private combinePaths(...paths: string[]): string {
-    return paths
-      .filter(Boolean)
-      .map(path => this.normalizePath(path))
-      .join('')
-      .replace(/\/+/g, '/');
-  }
-}
